refactor: LLM판결 오타 수정
