+++ src/main/java/store/lastdance/service/admin/AdminServiceImpl.java
@@ -24,7 +24,7 @@ import store.lastdance.dto.admin.stats.DashboardContentStats;
 import store.lastdance.dto.admin.stats.DashboardUserStats;
 import store.lastdance.dto.admin.stats.UserStats;
 import store.lastdance.dto.expense.ExpenseAnalysisHistoryDTO;
-import store.lastdance.repository.aijudement.AiJudgmentRepository;
+import store.lastdance.repository.aijudgment.AiJudgmentRepository;
 import store.lastdance.repository.expense.ExpenseAnalysisHistoryRepository;
 import store.lastdance.repository.user.UserRepository;
 import store.lastdance.repository.community.PostRepository;
